+++ src/main/kotlin/com/github/jensim/megamanipulator/settings/SerializationHolder.kt
@@ -1,5 +1,7 @@
 package com.github.jensim.megamanipulator.settings
 
+import com.fasterxml.jackson.annotation.JsonInclude
+import com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL
 import com.fasterxml.jackson.databind.DeserializationFeature.ACCEPT_EMPTY_ARRAY_AS_NULL_OBJECT
 import com.fasterxml.jackson.databind.DeserializationFeature.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT
 import com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES
@@ -13,11 +15,13 @@ object SerializationHolder {
         configure(FAIL_ON_UNKNOWN_PROPERTIES, false)
         configure(ACCEPT_EMPTY_STRING_AS_NULL_OBJECT, true)
         configure(ACCEPT_EMPTY_ARRAY_AS_NULL_OBJECT, true)
+        setDefaultPropertyInclusion(NON_NULL)
         registerModule(KotlinModule.Builder().build())
     }
 
     val objectMapper = new()
     val readable = new().apply {
         configure(INDENT_OUTPUT, true)
+        setDefaultPropertyInclusion(JsonInclude.Include.NON_DEFAULT)
     }
 }