+++ src/main/kotlin/com/github/jensim/megamanipulator/settings/types/codehost/CodeHostSettingsGroup.kt
@@ -12,6 +12,5 @@ data class CodeHostSettingsGroup(
         }
     }
 
-    val value: CodeHostSettings
-        get() = (gitHub ?: gitLab ?: bitBucket)!!
+    fun value(): CodeHostSettings = (gitHub ?: gitLab ?: bitBucket)!!
 }